requirements.txt
cognee
fastapi
uvicorn[standard]
python-dotenv